fixup! wip: create storage.Factory
