+++ pkg/state/factory.go
@@ -1,26 +1,25 @@
 package state
 
 import (
-	"github.com/filecoin-project/chain-validation/pkg/storage"
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
 )
 
-// Factory abstracts over concrete state manipulation methods.
-type Factory interface {
+// StateFactory abstracts over concrete state manipulation methods.
+type StateFactory interface {
 	NewActor(code cid.Cid, balance AttoFIL) Actor
 	NewState(actors map[Address]Actor) (Tree, error)
 	ApplyMessage(*VMContext, Tree, interface{}) (Tree, error)
 }
 
 type Validator struct {
-	factory Factory
-	store storage.StorageMap
+	factory StateFactory
+	store   StorageMap
 }
 
 
-func NewValidator(factory Factory, storage storage.Factory) *Validator {
+func NewValidator(factory StateFactory, storage StorageFactory) *Validator {
 	bs := blockstore.NewBlockstore(datastore.NewMapDatastore())
 	storageMap := storage.NewStorageMap(bs)
 
@@ -29,7 +28,7 @@ func NewValidator(factory Factory, storage storage.Factory) *Validator {
 
 // TODO probs make an interface
 type VMContext struct {
-	Store storage.StorageMap
+	Store StorageMap
 }
 
 func (v *Validator) ApplyMessages(tree Tree, messages []interface{}) (Tree, error) {
@@ -42,5 +41,17 @@ func (v *Validator) ApplyMessages(tree Tree, messages []interface{}) (Tree, erro
 		}
 	}
 	return tree, nil
+}
+type StorageFactory interface {
+	NewStorageMap(store blockstore.Blockstore) StorageMap
+}
+
+type StorageMap interface {
+	NewStorage(addr Address, actor Actor) Storage
+	Flush() error
+}
+
+type Storage interface {
+	Get(cid cid.Cid) ([]byte, error)
 }
